Move webdriver commands to common module.
